requirements.txt
Django==1.10.3
gunicorn==19.6.0
psycopg2==2.6.2
PyJWT
django-cors-headers
watson-developer-cloud==0.22.0
twilio==5.6.0
python-twitter==3.2
